Switch to component.
