+++ src/app/index.module.js
@@ -17,7 +17,7 @@ import {BrowseDatasetsController} from './browse/datasets/browse-datasets-contro
 import {BrowseTagsController} from './browse/tags/browse-tags-controller';
 import {BrowseAuthorsController} from './browse/authors/browse-author-controller';
 import {CommentDirective} from './details/comment-directive';
-import {DatasetDetailsOutlineDirective} from './details/mcpub-dataset-details-outline.directive';
+import './details/mcpub-dataset-details-outline.component';
 import {NavbarDirective} from '../app/components/navbar/navbar.directive';
 import {HomeTabDirective} from '../app/home/home-tab-directive';
 import {SearchBarDirective} from '../app/directives/search-bar-directive';
@@ -36,8 +36,9 @@ import {AccountsService} from './services/accounts-service.service';
 import {ValidateController} from './components/sign/validate/validate.controller';
 
 angular.module('mcpub', ['ngAnimate', 'ngCookies', 'ngTouch', 'ngSanitize', 'ngMessages', 'ngAria',
-        'restangular', 'ui.router', 'ui.bootstrap', 'toastr', 'ct.ui.router.extras', 'angularUtils.directives.dirPagination',
-        'ngTagsInput', 'ngFileUpload'])
+    'restangular', 'ui.router', 'ui.bootstrap', 'toastr', 'ct.ui.router.extras',
+    'angularUtils.directives.dirPagination',
+    'ngTagsInput', 'ngFileUpload'])
     .config(config)
     .config(routerConfig)
     .run(runBlock)
@@ -63,7 +64,6 @@ angular.module('mcpub', ['ngAnimate', 'ngCookies', 'ngTouch', 'ngSanitize', 'ngM
     .directive('loginDirective', LoginDirective)
     .directive('registerDirective', RegisterDirective)
     .directive('commentDirective', CommentDirective)
-    .directive('datasetDetailsOutline', DatasetDetailsOutlineDirective)
     .service('pubAPI', pubAPIService)
     .service('mcapi', mcapiService)
     .service('accountsService', AccountsService)
